Extract common mkmf code into nmatrix/mkmf.rb
